Allow to disable session identifier
In tasks, session identifier is used as a seed during the creation
of the task hash. The task hash determines the final process
working directory and makes the task unique among others.

This patch allows users to bypass that seed by setting the
`removeSessionIdentifier = true` property.

Example:

process {
    removeSessionIdentifier = true
}

This is intended for debugging/testing purposes at the moment.

Signed-off-by: Lorenzo Fontana <fontanalorenz@gmail.com>